Various tweaks, mostly to get LSP jedi-language-server working
Finally seem to have got a working LSP config to use jedi-language-server.

Needed to remove system install of python-language-server (pyls/pylsp) and the same that had been pip
install --user so that they were not taking priority over the user installed jedi-language-server (or
those under virtualenvs).

Moved custom-file to earlier in the config.

Improved config under magit with hook in correct place.

Set :defer 0.5 for a bunch of polymode packages